Ticket/stripe connect terms acceptance

+++ bluebottle/funding_stripe/views.py
@@ -121,12 +121,16 @@ class ConnectAccountDetails(JsonApiViewMixin, AutoPrefetchMixin, RetrieveUpdateA
     }
 
     def perform_update(self, serializer):
-        if serializer.instance.country != serializer.validated_data["country"]:
+        if (
+            "country" in serializer.validated_data
+            and serializer.instance.country != serializer.validated_data["country"]
+        ):
             if serializer.instance.status == "verified":
                 raise ValidationError("Cannot change country of verified account")
 
             serializer.instance.external_accounts.all().delete()
             serializer.instance.account_id = None
+            serializer.instance.tos_acceptance = False
 
         return super().perform_update(serializer)
